backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyJWT
passlib[bcrypt]
python-dotenv
requests